RTVfsIoStrmZeroFill: Fixed wrong method test (pfnSkip instead of pfnZeroFill); use g_abRTZero64K instead of the heap.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@67274 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/vfs/vfsbase.cpp
@@ -3050,7 +3050,7 @@ RTDECL(int) RTVfsIoStrmZeroFill(RTVFSIOSTREAM hVfsIos, RTFOFF cb)
         rc = VINF_SUCCESS;
         while (cb > 0)
         {
-            size_t cbToWrite = (size_t)RT_MIN(cb, sizeof(g_abRTZero64K));
+            size_t cbToWrite = (size_t)RT_MIN(cb, (ssize_t)sizeof(g_abRTZero64K));
             RTVfsLockAcquireWrite(pThis->Base.hLock);
             rc = RTVfsIoStrmWrite(hVfsIos, g_abRTZero64K, cbToWrite, true /*fBlocking*/, NULL);
             RTVfsLockReleaseWrite(pThis->Base.hLock);
